Improved portability and performance
* Improved portability for the portable build mode. * Boosted performance by enabling the -flto (Link Time Optimization) option.

+++ src/exit_status.c
@@ -33,5 +33,5 @@ int exit_status(
 		}
 	}
 
-	return(status);
+	return((int)status);
 }
